fix: fix full node evm e2e (#2669)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.

NOTE: PR titles should follow semantic commits:
https://www.conventionalcommits.org/en/v1.0.0/
-->

## Overview

The E2E test was stalling on the DAIncludedHeight due to a signing
verification error. We were signing the header instead of the data which
is required in the `da_retreiver`.

The height was also not being incremented when there was empty data. I
added additional logic to increment the DA height on successful DA
retrieval even if there are no events.

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 

Ex: Closes #<issue number>
-->

---------

Co-authored-by: Julien Robert <julien@rbrt.fr>