refactor: update evm tests to use tastora reth package (#2689)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.

NOTE: PR titles should follow semantic commits:
https://www.conventionalcommits.org/en/v1.0.0/
-->

## Overview

Currently, the evm tests use the testcontainers lib to do a docker
compose up with static ports.

With this PR, we replace the container provisioning logic with the reth
implementation in tastora (which is using the same containers).

Since tastora performs dynamic port allocation, we needed to update all
the startup commands to use the dynamically generated ports instead of
the static preset ones.

I didn't do it in this PR since it is an already large diff, but we
should be able to remove the docker compose files if they don't serve
another purpose after this is merged.

In a follow up we can also move the evmsingle logic to use the tastora
version or leave it as a local binary (or support both if that is
desirable.)

For now I just replaced the stuff that was already in docker with the
tastora implemenation.




<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 

Ex: Closes #<issue number>
-->