test: correct iteration over pids in SUT (#2693)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.

NOTE: PR titles should follow semantic commits:
https://www.conventionalcommits.org/en/v1.0.0/
-->

## Overview

NOTE: the fix was actually included in the e2e refactor, this PR is just
adding some extra tests to cover it.

I found a small bug in the `SystemUnderTest` type in which iteration was
happening over indices instead of process IDs, so when Shutdown was
called, it was not actually finding the process.

The fix was

```diff
- for pid := range pids {
+ for _, pid := range pids {
	p, err := os.FindProcess(pid)
	if err != nil {
		continue
	}
	if !yield(p) {
		break
	}
}
```

in `iterProcessesByCmd`


<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 

Ex: Closes #<issue number>
-->