refactor(syncing): remove retry da retriever (#2711)
## Overview

The da height is never bumped if a blob has not been fetched.
It is bumped at successful blob retrieval (see syncer): an empty blob or
a valid blobs bumps it.

Additionally we already have a backoff for height in the future in the
syncer.
This system has been applied to all errors, which makes retries faster,
as it waits for DA block time.

We should remove the retry logic within the retriever, as we do still
want to check p2p prior to that.
We should automatically retry already at the next loop iteration

-
https://github.com/evstack/ev-node/blob/v1.0.0-beta.5/block/internal/syncing/syncer.go#L281-L283
-
https://github.com/evstack/ev-node/blob/v1.0.0-beta.5/block/internal/syncing/syncer.go#L302-L303