fix: sync service for non zero height starts with empty store (#2834)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.

NOTE: PR titles should follow semantic commits:
https://www.conventionalcommits.org/en/v1.0.0/
-->

## Overview

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 

Ex: Closes #<issue number>
-->

sync service fix for when we are not on genesis but have an empty store


in 1.0.0-beta.9 we fixed writing to p2p store but the issue is when the
store is empty AND we are not in genesis then the store is not
initialised, this pr aims to fix this