feat: forced inclusion (#2797)
ref: https://github.com/evstack/ev-node/issues/1914

A choice has been made to make this logic in the executor and avoid
extending the reaper and the sequencer.
This is because, updating the repeer, means passing down the last
fetched da height accross all components.
It adds a lot of complexity otherwise. Adding it in the sequencer may be
preferable, but this makes the inclusion in a sync node less
straightforward. This is what is being investigated.

Compared to the previous implementation, a forced transaction does not
have any structure. It should be the raw structure from the execution
client. This is to keep ev-node know nothing about the transaction. No
signature checks, no validation of correctness. The execution client
must make sure to reject gibberish transactions.

- [x] implement for executor
  - [x] add size checks, to make sure batch doesn't go over max bytes
- [x] implement for syncer
 - requirements:
- [x] syncer must be able to detect if sequencer was malicious and
hasn't included a forced tx
     - [x] fetched inclusion of the txs bytes within the block data
     - [x] if that happened, reject all future sequencer blocks
       - [x] eventually, become based and keep produce block
- [x] tool for submitting forced txs in celestia easily ->
https://github.com/evstack/ev-abci/pull/295
- [x] make sure epochs are in sync between syncer and executor


---- for later, won't be included in this pr (ref
https://github.com/evstack/ev-node/pull/2797#issuecomment-3512654215)

- [ ] improve offline capabilities (if sequencer goes offline, full node
become fully based)
- [ ] when sequencer comes back only, needs to fetch if any block was
produced :/

---------

Co-authored-by: Marko <marko@baricevic.me>