feat: DA Client remove interface part 2,  add client for celestia blob api   (#2909)
Parent Epic: #2796 

- In order to advance the removing of core/da the next step is to create
a client that won't be wired in this PR that consumes the new blob API.

## New client (namespace blob) - New client not yet wired
- Json RPC: **da/jsonrpc/blob/client.go** opens JSON-RPC to Celestia’s
blob namespace.
- Business wrapper: **block/internal/da/blob_client.go** wraps the
jsonrpc to add additional logic for our business logic, some things that
were done on the old da namespace is not done anymore and is expected to
be done by the client.

Both clients coexist for now (old and new), but the new one is not
wired.

The idea is to wire the New client in another PR, because it is the
biggest part and it will generate a lot of code changes.

<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.

NOTE: PR titles should follow semantic commits:
https://www.conventionalcommits.org/en/v1.0.0/
-->

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 

Ex: Closes #<issue number>
-->