fix(block): fix init logic sequencer for da epoch fetching (#2926)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.

NOTE: PR titles should follow semantic commits:
https://www.conventionalcommits.org/en/v1.0.0/
-->

## Overview

The sequencer doesn't know at genesis what will be the first da block it
publishes to.
So it is a manual operation to edit the genesis and share it with other
nodes afterwards.
We need the sequencer to start with a pre-init da height, as we use the
daHeight from the genesis as well.
Basically it should be the current da height at the time the first block
of the chain was created.
For epochs to work, this would need to be the same across all nodes (so
you CANNOT change it).

This PR let the first da height be set based on the first submitted
height for a single sequencer
For a based sequencer, the chain developers basically need to have a da
height set in the genesis.
Added a check to avoid issues.


ref: https://github.com/evstack/ev-node/issues/2906

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 

Ex: Closes #<issue number>
-->

---------

Co-authored-by: Marko <marko@baricevic.me>